MEXSVR-410 将来始まる軽減税率を見据えて、各種テーブルの税率のカラム(TaxRate)のデータ型を揃える
git-svn-id: http://svnsin.ncr.com/svn/repos/pa/res/RES%20MeX/Base/resTransaction/branches/3.3.2@76845 9234da4d-b7fe-4b44-acf3-10f21a033f33

+++ src/ncr/res/mobilepos/pricing/helper/ItemHelper.java
@@ -89,7 +89,7 @@ public final class ItemHelper {
      * @param rounding
      * @return
      */
-	private static long getSalesPriceWithTaxRounding(long salesprice, int taxRate,
+	private static long getSalesPriceWithTaxRounding(long salesprice, double taxRate,
 			int rounding) {
 
 		double salespricewithtaxAmount = salesprice + salesprice
@@ -113,7 +113,7 @@ public final class ItemHelper {
 	 * @param rounding
 	 * @return
 	 */
-	private static long getSalesUnitPriceWithTaxRounding(long regSalesUnitPrice, int taxRate,
+	private static long getSalesUnitPriceWithTaxRounding(long regSalesUnitPrice, double taxRate,
 			int rounding) {
 
 		double salesprice =
